# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Shop API
  version: 1.0.0

paths:
  /api/users:
    post:
      tags:
        - Users
      summary: Создать пользователя
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    get:
      tags:
        - Users
      summary: Получить всех пользователей
      operationId: getAllUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Получить пользователя по id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found

    put:
      tags:
        - Users
      summary: Обновить пользователя
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Users
      summary: Удалить пользователя
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

  /api/categories:
    post:
      tags:
        - Categories
      summary: Создать категорию
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
    get:
      tags:
        - Categories
      summary: Получить все категории
      operationId: getAllCategories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'

  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Получить категорию по id
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Not found

    put:
      tags:
        - Categories
      summary: Обновить категорию
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Categories
      summary: Удалить категорию
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

  /api/products:
    post:
      tags:
        - Products
      summary: Создать товар
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    get:
      tags:
        - Products
      summary: Получить все товары
      operationId: getAllProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Получить товар по id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found

    put:
      tags:
        - Products
      summary: Обновить товар
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Products
      summary: Удалить товар
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

  /api/carts:
    post:
      tags:
        - Carts
      summary: Создать корзину
      operationId: createCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
    get:
      tags:
        - Carts
      summary: Получить все корзины
      operationId: getAllCarts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartResponse'

  /api/carts/{id}:
    get:
      tags:
        - Carts
      summary: Получить корзину по id
      operationId: getCartById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Carts
      summary: Удалить корзину
      operationId: deleteCart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

  /api/orders:
    post:
      tags:
        - Orders
      summary: Создать заказ
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    get:
      tags:
        - Orders
      summary: Получить все заказы
      operationId: getAllOrders
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Получить заказ по id
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Orders
      summary: Удалить заказ
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

  /api/deliveries:
    post:
      tags:
        - Deliveries
      summary: Создать доставку
      operationId: createDelivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeliveryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryResponse'
    get:
      tags:
        - Deliveries
      summary: Получить все доставки
      operationId: getAllDeliveries
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeliveryResponse'

  /api/deliveries/{id}:
    get:
      tags:
        - Deliveries
      summary: Получить доставку по id
      operationId: getDeliveryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryResponse'
        '404':
          description: Not found

    put:
      tags:
        - Deliveries
      summary: Обновить доставку
      operationId: updateDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeliveryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryResponse'
        '404':
          description: Not found

    delete:
      tags:
        - Deliveries
      summary: Удалить доставку
      operationId: deleteDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
        '404':
          description: Not found

components:
  schemas:
    UserRequest:
      type: object
      properties:
        username:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string

    CategoryRequest:
      type: object
      properties:
        name:
          type: string

    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    ProductRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        categoryId:
          type: string
          format: uuid

    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
        category:
          $ref: '#/components/schemas/CategoryResponse'

    CartRequest:
      type: object
      properties:
        productIds:
          type: array
          items:
            type: string
            format: uuid

    CartResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'

    OrderRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        delivery:
          $ref: '#/components/schemas/DeliveryRequest'

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        delivery:
          $ref: '#/components/schemas/DeliveryResponse'

    DeliveryRequest:
      type: object
      properties:
        address:
          type: string

    DeliveryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string